// SNMP_FRONT/src/app/model/names.olts.ts
export interface Olts {
    id:               string;
    id_olt:           number;
    nombre:           string;
    ip:               string;
    descricion:       null;
    tecnologia:       string;
    id_region:        number;
    id_configuracion: number;
    estatus:          number;
    _id:              string;
}